Code review

+++ src/BackendBehaviors.php
@@ -46,14 +46,14 @@ class BackendBehaviors
                 $ret .= '<li class="line" id="dmsp' . $rs->post_id . '">';
                 $ret .= '<a href="' . App::backend()->url()->get('admin.post', ['id' => $rs->post_id]) . '">' . $rs->post_title . '</a>';
                 if ($large) {
-                    $dt = '<time datetime="' . Date::iso8601(strtotime($rs->post_dt), App::auth()->getInfo('user_tz')) . '">%s</time>';
+                    $dt = '<time datetime="' . Date::iso8601((int) strtotime($rs->post_dt), App::auth()->getInfo('user_tz')) . '">%s</time>';
                     $ret .= ' (' .
                     __('by') . ' ' . $rs->user_id . ' ' . sprintf($dt, __('on') . ' ' .
                         Date::dt2str(App::blog()->settings()->system->date_format, $rs->post_dt) . ' ' .
                         Date::dt2str(App::blog()->settings()->system->time_format, $rs->post_dt)) .
                     ')';
                 } else {
-                    $ret .= ' (<time datetime="' . Date::iso8601(strtotime($rs->post_dt)) . '">' . Date::dt2str(__('%Y-%m-%d %H:%M'), $rs->post_dt) . '</time>)';
+                    $ret .= ' (<time datetime="' . Date::iso8601((int) strtotime($rs->post_dt)) . '">' . Date::dt2str(__('%Y-%m-%d %H:%M'), $rs->post_dt) . '</time>)';
                 }
                 $ret .= '</li>';
             }
